Renamed GMatLinearElastic -> GMatElastic. Simplified implementation, using Elastic class.
