+++ GMatElastic.pc.in
