+++ GMatElasticConfig.cmake.in
