+++ setup.py
@@ -9,16 +9,16 @@ import setuptools
 import pybind11
 import pyxtensor
 
-header = open('include/GMatLinearElastic/config.h','r').read()
-world  = re.split(r'(.*)(\#define GMATLINEARELASTIC_WORLD_VERSION\ )([0-9]+)(.*)',header)[3]
-major  = re.split(r'(.*)(\#define GMATLINEARELASTIC_MAJOR_VERSION\ )([0-9]+)(.*)',header)[3]
-minor  = re.split(r'(.*)(\#define GMATLINEARELASTIC_MINOR_VERSION\ )([0-9]+)(.*)',header)[3]
+header = open('include/GMatElastic/config.h','r').read()
+world  = re.split(r'(.*)(\#define GMATELASTIC_WORLD_VERSION\ )([0-9]+)(.*)',header)[3]
+major  = re.split(r'(.*)(\#define GMATELASTIC_MAJOR_VERSION\ )([0-9]+)(.*)',header)[3]
+minor  = re.split(r'(.*)(\#define GMATELASTIC_MINOR_VERSION\ )([0-9]+)(.*)',header)[3]
 
 __version__ = '.'.join([world,major,minor])
 
 ext_modules = [
   Extension(
-    'GMatLinearElastic',
+    'GMatElastic',
     ['python/main.cpp'],
     include_dirs=[
       pybind11.get_include(False),
@@ -33,14 +33,14 @@ ext_modules = [
 ]
 
 setup(
-  name             = 'GMatLinearElastic',
+  name             = 'GMatElastic',
   description      = 'Linear elastic material model',
   long_description = desc,
   version          = __version__,
   license          = 'MIT',
   author           = 'Tom de Geus',
   author_email     = 'tom@geus.me',
-  url              = 'https://github.com/tdegeus/GMatLinearElastic',
+  url              = 'https://github.com/tdegeus/GMatElastic',
   ext_modules      = ext_modules,
   install_requires = ['pybind11>=2.2.0'],
   cmdclass         = {'build_ext': pyxtensor.BuildExt},
